requirements.txt
requests
beautifulsoup4
GitPython
pymisp